ai-core/requirements.txt
matplotlib==3.5.1
numpy==1.21.5
pandas==1.3.5
pyforest==1.0.2
scipy==1.9.2
sklearn
pandas 
dash
httpx==0.20 
streamlit 
plotly
Flask
sentence-transformers==2.2.2
PyPDF2==3.0.1
torch==2.0.0
reportlab==3.6.12
python-dotenv==1.0.0
python-docx==0.8.11
langdetect==1.0.9
nltk==3.8.1
